fuzz/requirements.txt
eth-ape>=0.7.0
ape-vyper
ape-anvil
hypothesis
pytest
python-dotenv
decimalfp
